+++ test/lib/odeon_uk/screening_test.rb
@@ -1,5 +1,0 @@
-require_relative '../../test_helper'
-
-describe OdeonUk::Screening do
-  let(:described_class) { OdeonUk::Screening }
-end